+++ hybrid_core/hybrid_core_windows/lib/src/ffi.dart
@@ -5,7 +5,7 @@ import 'dart:ffi';
 import 'ffi.g.dart';
 
 final _kernel32Lib = DynamicLibrary.open('Kernel32.dll');
-final kernel32Lib = WindowsCoreLibrary(_kernel32Lib);
+final kernel32Lib = HybridCoreWindowsLibrary(_kernel32Lib);
 
 int LOBYTE(int w) => w & 0xff;
 int HIBYTE(int w) => (w >> 8) & 0xff;